Using Canvas to display Woka in Woka select screens
Using a div with a background color caused images to be loaded and put in cache without CORS headers.
This is an issue with S3 bucket that store images even when CORS is configured.
As a result, when entering the map, Phaser would load in AJAX the image that is in cache, without CORS. It would not work.
By loading everything from scratch in AJAX, we avoid this problem.
