+++ play/src/front/Components/Woka/WokaPreview.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
-    import { onMount, onDestroy, createEventDispatcher } from "svelte";
-    import { WokaData, WokaTexture } from "./WokaTypes";
+    import { createEventDispatcher } from "svelte";
+    import { WokaData } from "./WokaTypes";
+    import WokaImage from "./WokaImage.svelte";
 
     const dispatch = createEventDispatcher<{ rotate: { direction: number } }>();
 
@@ -8,78 +9,12 @@
     export let wokaData: WokaData | null = null;
     export let getTextureUrl: (url: string) => string = (url) => url;
 
-    const bodyPartOrder = ["body", "eyes", "hair", "clothes", "hat", "accessory", "woka"];
-
     // Directions correspond to the order of images in the sprite sheet:
-    const directionsCoresp = [0, 1, 3, 2];
+    const directionsMapping = [0, 1, 3, 2];
 
-    let canvas: HTMLCanvasElement;
-    let ctx: CanvasRenderingContext2D;
-    let images: Record<string, HTMLImageElement> = {};
-    let frame: number = 0;
     let direction: number = 0;
-    let raf: number;
-    let frameCount: number = 0;
-    const framesPerStep = 15;
-    const canvaSize = 130;
+    const canvasSize = 130;
 
-    function findTextureUrl(bodyPart: string): string | null {
-        const textureId = selectedTextures?.[bodyPart];
-        if (!textureId || !wokaData?.[bodyPart]?.collections) return null;
-        for (const collection of wokaData[bodyPart].collections) {
-            const texture = collection.textures.find((t: WokaTexture) => t.id === textureId);
-            if (texture) return getTextureUrl(texture.url);
-        }
-        return null;
-    }
-
-    function loadImages() {
-        images = {};
-        for (const part of bodyPartOrder) {
-            const url = findTextureUrl(part);
-            if (url) {
-                const img = new window.Image();
-                img.src = url;
-                images[part] = img;
-            }
-        }
-    }
-
-    function draw() {
-        ctx.clearRect(0, 0, canvas.width, canvas.height);
-        for (const part of bodyPartOrder) {
-            const img = images[part];
-            ctx.imageSmoothingEnabled = false;
-            if (img && img.complete) {
-                ctx.drawImage(img, frame * 32, directionsCoresp[direction] * 32, 32, 32, 0, 0, canvaSize, canvaSize);
-            }
-        }
-    }
-
-    function animate() {
-        frameCount++;
-        if (frameCount >= framesPerStep) {
-            frame = (frame + 1) % 3;
-            frameCount = 0;
-        }
-        draw();
-        raf = requestAnimationFrame(animate);
-    }
-
-    $: if (selectedTextures) {
-        loadImages();
-    }
-
-    onMount(() => {
-        const context = canvas.getContext("2d");
-        if (!context) return;
-        ctx = context;
-        loadImages();
-        animate();
-    });
-    onDestroy(() => {
-        cancelAnimationFrame(raf);
-    });
 </script>
 
 <div class="woka-preview flex items-center justify-center relative">
@@ -88,7 +23,7 @@
             class="bg-white/10 hover:bg-white/20 rounded-md absolute bottom-2 right-2 aspect-square p-2 flex items-center justify-center"
             on:click={() => {
                 direction = (direction + 1) % 4;
-                dispatch("rotate", { direction: directionsCoresp[direction] });
+                dispatch("rotate", { direction: directionsMapping[direction] });
             }}
         >
             <svg
@@ -107,12 +42,6 @@
                 /><path d="M20 4v5h-5" /></svg
             >
         </button>
-        <canvas
-            bind:this={canvas}
-            width={canvaSize}
-            height={canvaSize}
-            style="image-rendering: pixelated;"
-            class="z-500"
-        />
+        <WokaImage {selectedTextures} {wokaData} {getTextureUrl} {canvasSize} direction={directionsMapping[direction]} />
     </div>
 </div>
